+++ examples/web/src/components/PageLink.vue
@@ -3,11 +3,14 @@ import { Card, CardContent, CardHeader } from '@scalar/api-reference'
 import { RouterLink } from 'vue-router'
 
 defineProps<{
-  to: string
+  to?: string
+  href?: string
 }>()
 </script>
 <template>
-  <RouterLink :to="{ name: to }">
+  <RouterLink
+    v-if="to"
+    :to="{ name: to }">
     <Card>
       <CardHeader>
         <slot name="title" />
@@ -17,6 +20,18 @@ defineProps<{
       </CardContent>
     </Card>
   </RouterLink>
+  <a
+    v-else-if="href"
+    :href="href">
+    <Card>
+      <CardHeader>
+        <slot name="title" />
+      </CardHeader>
+      <CardContent class="description">
+        <slot name="description" />
+      </CardContent>
+    </Card>
+  </a>
 </template>
 
 <style scoped>