+++ examples/web/src/pages/SwaggerEditorPage.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-import { SwaggerEditor } from '@scalar/swagger-editor'
-</script>
-
-<template>
-  <SwaggerEditor />
-</template>