+++ examples/web/vite.config.ts
@@ -23,7 +23,7 @@ export default defineConfig({
     }),
   ],
   server: {
-    port: 5050,
+    port: 3000,
   },
   resolve: {
     alias: [