+++ packages/api-client-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @scalar/api-client-react
 
+## 0.1.4
+
+### Patch Changes
+
+- c200170: fix: react client props name
+
+## 0.1.3
+
+### Patch Changes
+
+- d91133e: feat: add proxy as prop to react client
+
 ## 0.1.2
 
 ### Patch Changes